requirements.txt
python-bsonjs==0.4.0